ml_service/requirements.txt
scikit-learn>=1.0.2
pandas>=1.3.0
numpy>=1.21.0
tweepy==4.14.0
python-dotenv==1.0.1
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
joblib==1.2.0
nltk==3.8.1
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0 